Models added for the Throne Room
The pillars added to the throne room (unsure on the roof height so it was made to be adjustable for any feasible height). A nice fancy torch for decoration around the throne room but could be implemented across any other room. Reference Person made as a seperate model for easy incorporation into any Maya scene without the controllers attached.