More furnishing models complete
The flower pot (no plant inside the pot), and the chandelier (no torches) have been created. The plant is not present to allow for versatility in what goes in the pot, and no candles on the chandelier to allow for candle model to be created and used within the chandelier itself for reusability sakes.